consumer/requirements.txt
cassandra-driver==3.25.0
kafka-python==2.0.2
pandas==1.4.2